refactored passenger and other minor changes
