+++ app/available-rides/page.jsx
@@ -250,4 +250,3 @@ return (
 };
 
 export default AvailableRides;
-
