+++ app/styles/available-rides.module.css
